+++ app.py
@@ -1,4 +1,4 @@
 # from canvas_tutorial.canvasTutorial import main
-from youtube_tutorial.youtubeTutorial import main
+from youtube_tutorial.basicQLearning import main
 
 main('MountainCar-v0')
